api-gateway/requirements.txt
# API Gateway - Independent Dependencies
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# HTTP Client (for routing to microservices)
httpx==0.27.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utils
python-dotenv==1.0.1
loguru==0.7.2

# Rate limiting
slowapi==0.1.9

# Development & Testing
pytest==8.3.0
pytest-asyncio==0.23.0
